requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bcrypt==4.2.0
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docutils==0.19
fastapi==0.109.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown2==2.5.0
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==8.4.0
nicegui==1.4.26
orjson==3.10.7
pscript==0.7.7
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
schedule==1.2.2
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vbuild==0.8.2
watchfiles==0.24.0
websockets==13.0.1
wsproto==1.2.0
yarl==1.10.0